requirements..txt
# Entorno y configuración
python-dotenv==1.1.1

# Datos
pandas==2.3.2
PyYAML==6.0
Faker==37.6.0

# LLM / Agentes
langchain==0.3.27
langchain-core==0.3.75
langchain-community==0.3.27
langchain-openai==0.3.27
tiktoken==0.11.0
openai==1.106.1
langgraph==0.6.6

# Base de datos
SQLAlchemy==2.0.41

# Otros útiles
requests==2.32.4